Add \n character before sending to tempfile.do
Adding the newline character ensures the file runs, even when there's only one line sent. 
I was having an issue where I couldn't send the line I had my cursor on without fully selecting it. Even then it would fail to run sometimes. I'm guessing it's because Stata requires an end-of-line character to run the line (which didn't exist when there's only one line selected).  This seems to have fixed it for me.